third_party/blink/web_tests/wpt_internal/bluetooth/characteristic/notifications/start-succeeds.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithTrustedClick({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('heart_rate'))
    .then(service => service.getCharacteristic('heart_rate_measurement'))
    .then(characteristic => {
      return characteristic.startNotifications()
        .then(start_characteristic => {
          assert_equals(start_characteristic, characteristic,
                  'Start characteristic should be the same as characteristic.');
        });
    });
  // TODO(ortuno): Assert that notifications are active.
  // http://crbug.com/600762
}, 'Single start notifications succeeds.');
</script>
